컨트롤러 및 요청 dto 클래스명 변경을 위한 사전 작업
